refactor: move base popup logic from Popup to PopupWrapper
