+++ src/components/Template/Popup/Popup.module.scss
@@ -1,35 +1,15 @@
 @use '../../../globalStyle/variables' as v;
 
 .popup {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 98;
-
-  &:before {
-    position: absolute;
-    top: 0;
-    left: 0;
-    content: '';
-    width: 100%;
-    height: 100%;
-    background: v.$black-color;
-    opacity: .5;
-  }
-
-  &__container {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 100%;
-    min-height: 150px;
-    padding: 32px 48px;
-    border-radius: 16px;
-    box-shadow: 0 4px 24px 0 rgba(0, 0, 0, 0.08);
-    background: v.$white-color;
-    z-index: 99;
-  }
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 100%;
+  min-height: 150px;
+  padding: 32px 48px;
+  border-radius: 16px;
+  box-shadow: 0 4px 24px 0 rgba(0, 0, 0, 0.08);
+  background: v.$white-color;
+  z-index: 99;
 }
